Move taxonomy form to its own file
